requirements.txt
aiogram~=2.19
aioredis
gspread
oauth2client~=4.1.3
google-api-python-client~=2.44.0
google-auth-httplib2
google-auth-oauthlib

gspread-asyncio
gspread-formatting
Faker
SQLAlchemy~=1.4.32
alembic~=1.7.7
httplib2~=0.20.4